Added a git ignore
A .gitignore file specifies which files to track or not track in a git repository. This gitignore file tells git to not track any new visual studio files that come around because many of these files are generated by visual studio and make tracking the code base by git more complicated.